Change input to int
To trigger DividebyZero properly

+++ Activities/Activity10_ExceptionHandling/Program.cs
@@ -5,7 +5,7 @@ Console.Write("Enter the first number: ");
 string? firstInput = Console.ReadLine();
 
 // Try converting to a number
-if (!double.TryParse(firstInput, out double num1))
+if (!int.TryParse(firstInput, out int num1))
 {
     Console.WriteLine("Invalid input. Please enter a valid number.");
     return;
@@ -17,7 +17,7 @@ Console.WriteLine($"First number accepted: {num1}");
 Console.Write("Enter the second number: ");
 string? secondInput = Console.ReadLine();
 
-if (!double.TryParse(secondInput, out double num2))
+if (!int.TryParse(secondInput, out int num2))
 {
     Console.WriteLine("Invalid input. Please enter a valid number.");
     return;
@@ -29,7 +29,7 @@ Console.WriteLine($"Numbers entered: {num1} and {num2}");
 // Divide the numbers with exception handling
 try
 {
-    double result = num1 / num2;
+    int result = num1 / num2;
     Console.WriteLine($"{num1} divided by {num2} equals {result}");
 }
 catch (DivideByZeroException)
